git-svn-id: https://adam.cs.princeton.edu/svn/craq/src@2565 41643c41-6115-4a67-97d8-b2dbeab842be

+++ donar/donar_rpc.h
@@ -10,7 +10,7 @@
 struct traffic_entry {
   u_int32_t server_id;
   u_int32_t update_time;
-  u_int32_t traffic_vol;
+  uint64_t traffic_vol;
 };
 void *traffic_entry_alloc ();
 bool_t xdr_traffic_entry (XDR *, void *);
